refactor(weather): break out finding location
With this change the location of the user will be cached for the
duration of the whole process. In the vast majority of cases, this is
fine. There is an edge case though that for people who are actively
traveling while using autolocate the location will not update. A
future task would be to invalidate the cache on a timer, or to possibly
separate the autolocate from the rest of `find_location` so that the
cache can be invalidated if the result from the autolocate service changes.

The added location struct also caches the name of the location. This was
previously fetched during the call to /weather, but the /onecall
endpoint does not contain this data.